Created GoogleCalendar Component, refactored Events calendar based on the Component, and added Calendar light Green from Figma to RTCColors
